+++ src/client/rpcserver.cpp
@@ -28,8 +28,8 @@ CryptoServer::CryptoServer(jsonrpc::AbstractServerConnector &connector) : Crypto
 }
 
 void CryptoServer::setWallet(CryptoKernel::Wallet* Wallet,
-                             CryptoKernel::Blockchain* Blockchain, 
-                             CryptoKernel::Network* Network, 
+                             CryptoKernel::Blockchain* Blockchain,
+                             CryptoKernel::Network* Network,
                              bool* running) {
     wallet = Wallet;
     blockchain = Blockchain;
@@ -46,16 +46,15 @@ Json::Value CryptoServer::getinfo() {
     std::stringstream buffer;
     buffer << std::setprecision(8) << std::fixed << balance;
     returning["balance"] = buffer.str();
-    returning["height"] = static_cast<unsigned long long int>
-                          (blockchain->getBlockDB("tip").getHeight());
+    returning["height"] = blockchain->getBlockDB("tip").getHeight();
     returning["connections"] = network->getConnections();
     returning["mempool"]["count"] = blockchain->mempoolCount();
-    
+
     buffer.str("");
-    buffer << std::setprecision(3) 
-           << (blockchain->mempoolSize() / double(1024 * 1024)) 
+    buffer << std::setprecision(3)
+           << (blockchain->mempoolSize() / double(1024 * 1024))
            << " MB";
-    
+
     returning["mempool"]["size"] = buffer.str();
 
     return returning;
@@ -74,8 +73,8 @@ Json::Value CryptoServer::account(const std::string& account,
         } catch(const CryptoKernel::Wallet::WalletException& e) {
             return Json::Value(e.what());
         }
-    }   
-    
+    }
+
     const auto newAccount = CryptoKernel::Wallet::Account(accJson);
 
     returning["name"] = newAccount.getName();
@@ -94,7 +93,7 @@ Json::Value CryptoServer::account(const std::string& account,
     return returning;
 }
 
-std::string CryptoServer::sendtoaddress(const std::string& address, double amount, 
+std::string CryptoServer::sendtoaddress(const std::string& address, double amount,
                                         const std::string& password) {
     uint64_t Amount = amount * 100000000;
     return wallet->sendToAddress(address, Amount, password);
@@ -249,7 +248,7 @@ Json::Value CryptoServer::importprivkey(const std::string& name, const std::stri
         return wallet->importPrivKey(name, key, password).toJson();
     } catch(const CryptoKernel::Wallet::WalletException& e) {
         return Json::Value(e.what());
-    } 
+    }
 }
 
 Json::Value CryptoServer::getpeerinfo() {
@@ -259,15 +258,11 @@ Json::Value CryptoServer::getpeerinfo() {
         Json::Value stat;
         stat["ping"] = stats.second.ping;
         stat["incoming"] = stats.second.incoming;
-        stat["connectedSince"] = static_cast<unsigned long long int>
-                                 (stats.second.connectedSince);
-        stat["transferUp"] = static_cast<unsigned long long int>
-                             (stats.second.transferUp);
-        stat["transferDown"] = static_cast<unsigned long long int>
-                               (stats.second.transferDown);
+        stat["connectedSince"] = stats.second.connectedSince;
+        stat["transferUp"] = stats.second.transferUp;
+        stat["transferDown"] = stats.second.transferDown;
         stat["version"] = stats.second.version;
-        stat["height"] = static_cast<unsigned long long int>
-                         (stats.second.blockHeight);
+        stat["height"] = stats.second.blockHeight;
         returning[stats.first] = stat;
     }
 
@@ -286,7 +281,7 @@ Json::Value CryptoServer::dumpprivkeys(const std::string& account,
                 return Json::Value(e.what());
             }
         }
-        
+
         return returning;
     } catch(const CryptoKernel::Wallet::WalletException& e) {
         return Json::Value(e.what());
